moved ProjectView files under the corresponding folder
